requirements.txt
numba~=0.48
python_speech_features
matplotlib~=3.2.1
torch~=1.4.0
pandas~=1.0.4
scikit-learn~=0.23.0
tqdm~=4.46.1
numpy~=1.19.0
Pillow~=7.1.2
torchvision~=0.5.0
nltk~=3.5
spacy~=2.2.4
gensim~=3.8.3
scipy~=1.4.1
transformers~=2.11.0
librosa~=0.7.2
torchaudio~=0.4.0
pydub~=0.24.1
PyYAML~=5.3.1